backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
pgvector>=0.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
slowapi==0.1.9
fakeredis==2.23.2
structlog>=24.0.0
prometheus-client>=0.19.0
# Generation pipeline dependencies
google-genai>=1.0.0
jinja2>=3.0.0
playwright>=1.40.0
pypdf>=4.0.0